Move check_encode_decode_invariant to test_utils
