+++ src/rust/rust-proto-new/tests/test_utils/mod.rs
@@ -1,1 +1,2 @@
+pub mod serde;
 pub mod strategies;
