+++ src/rust/rust-proto-new/tests/test_utils/serde.rs
@@ -1,0 +1,15 @@
+use std::fmt::Debug;
+
+use rust_proto_new::SerDe;
+
+// helper function to define a simple encode-decode invariant
+// see: https://hypothesis.works/articles/encode-decode-invariant/
+pub fn check_encode_decode_invariant<T>(serializable: T)
+where
+    T: SerDe + PartialEq + Clone + Debug,
+{
+    let cloned = serializable.clone();
+    let serialized = serializable.serialize().expect("serialization failed");
+    let deserialized = T::deserialize(serialized).expect("deserialization failed");
+    assert!(cloned == deserialized);
+}
